backend/requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
pymongo==4.6.1
flask-cors==4.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
bson==0.5.10
DNSPython==2.6.0
requests==2.31.0
pyjwt==2.8.0
passlib==1.7.4
loguru==0.7.2
flask-jwt-extended==4.6.0
flask-limiter==3.5.0
# Sentiment analysis dependencies
transformers==4.35.2
torch==2.1.1
# Text summarization dependencies
google-generativeai==0.3.1